change the ID of Message

+++ internal/service/message_service/message.go
@@ -20,11 +20,11 @@ var messageDao = entity.GetMessageDaoInstance()
 // SendMessage
 // @description 消息发送服务，将接收到的消息保存到MySQL数据库 TODO: 考虑要不要加入MQ
 // @auth	hqy			2023/08/17
-// @param	fromUserId	uint	发送方用户Id
-// @param	toUserId	uint	接收方用户Id
+// @param	fromUserId	int64	发送方用户Id
+// @param	toUserId	int64	接收方用户Id
 // @param	content		string	消息内容
 // @return	err			error	当执行出现错误时返回error，否则返回nil
-func SendMessage(fromUserId uint, toUserId uint, content string) error {
+func SendMessage(fromUserId int64, toUserId int64, content string) error {
 	// 验证toUserId的合法性
 	_, err := userDao.QueryUserById(int64(toUserId))
 	if err != nil {
@@ -41,11 +41,11 @@ func SendMessage(fromUserId uint, toUserId uint, content string) error {
 // QueryMessage
 // @description 查询某两个用户消息列表，转换为目标格式输出
 // @auth	hqy				2023/08/17
-// @param	fromUserId		uint			发送方用户Id
-// @param	toUserId		uint			接收方用户Id
+// @param	fromUserId		int64			发送方用户Id
+// @param	toUserId		int64			接收方用户Id
 // @return	messageInfoList	[]*MessageInfo	需要将Id转换为int64类型，并转换时间格式
 // @return	err				error			当执行出现错误时返回error，否则返回nil
-func QueryMessage(fromUserId uint, toUserId uint) ([]*MessageInfo, error) {
+func QueryMessage(fromUserId int64, toUserId int64) ([]*MessageInfo, error) {
 	// 从DAO层查询消息列表
 	messageList, err := messageDao.QueryMessage(fromUserId, toUserId)
 	if err != nil {
@@ -68,8 +68,8 @@ func QueryMessage(fromUserId uint, toUserId uint) ([]*MessageInfo, error) {
 
 // QueryLastMessage
 // @description 查询某两个用户最新一条消息，转换为目标格式输出
-// @param	fromUserId		uint			发送方用户Id
-// @param	toUserId		uint			接收方用户Id
+// @param	fromUserId		int64			发送方用户Id
+// @param	toUserId		int64			接收方用户Id
 // @return	messageInfo	*MessageInfo		需要将Id转换为int64类型，并转换时间格式
 // @return	err				error			当执行出现错误时返回error，否则返回nil
 func QueryLastMessage(fromUserId int64, toUserId int64) (*MessageInfo, error) {
